perf benchmark improvements
- allow running locally using `--dry-run`.
- limit CI runs to `main` branch only, so that we don't pollute the dashboard with merge queue runs. We can still trigger manually for other branches if needed.
